Move f142 FlatbufferReader implementation to separate compilation unit
